+++ src/utils/mod.rs
@@ -1,5 +1,19 @@
 use std::fmt::Write;
 
+use crate::{
+    backends::Backend,
+    constants::Span,
+    error::{ErrorKind, Result},
+    helpers::PrettyField,
+    mast::Mast,
+    parser::types::{ModulePath, TyKind},
+    type_checker::FullyQualified,
+    var::ConstOrCell,
+    witness::WitnessEnv,
+};
+
+use num_traits::One;
+
 pub fn noname_version() -> String {
     format!("@ noname.{}\n", env!("CARGO_PKG_VERSION"))
 }
@@ -85,4 +99,187 @@ yz
             (2, 5, "efgh\nijkl")
         );
     }
+}
+
+// logging utils
+pub fn log_array_type<B: Backend>(
+    backend: &B,
+    var_info_var: &[ConstOrCell<B::Field, B::Var>],
+    base_type: &TyKind,
+    size: u32,
+    witness: &mut WitnessEnv<B::Field>,
+    typed: &Mast<B>,
+    span: &Span,
+) -> Result<(String, Vec<ConstOrCell<B::Field, B::Var>>)> {
+    match base_type {
+        TyKind::Field { .. } => {
+            let values: Vec<String> = var_info_var
+                .iter()
+                .take(size as usize)
+                .map(|cvar| match cvar {
+                    ConstOrCell::Const(cst) => cst.pretty(),
+                    ConstOrCell::Cell(cell) => backend.compute_var(witness, cell).unwrap().pretty(),
+                })
+                .collect();
+
+            let remaining = var_info_var[size as usize..].to_vec();
+            Ok((format!("[{}]", values.join(", ")), remaining))
+        }
+
+        TyKind::Bool => {
+            let values: Vec<String> = var_info_var
+                .iter()
+                .take(size as usize)
+                .map(|cvar| match cvar {
+                    ConstOrCell::Const(cst) => {
+                        let val = *cst == B::Field::one();
+                        val.to_string()
+                    }
+                    ConstOrCell::Cell(cell) => {
+                        let val = backend.compute_var(witness, cell).unwrap() == B::Field::one();
+                        val.to_string()
+                    }
+                })
+                .collect();
+
+            let remaining = var_info_var[size as usize..].to_vec();
+            Ok((format!("[{}]", values.join(", ")), remaining))
+        }
+
+        TyKind::Array(inner_type, inner_size) => {
+            let mut nested_result = Vec::new();
+            let mut remaining = var_info_var.to_vec();
+            for _ in 0..size {
+                let (chunk_result, new_remaining) = log_array_type(
+                    backend,
+                    &remaining,
+                    inner_type,
+                    *inner_size,
+                    witness,
+                    typed,
+                    span,
+                )?;
+                nested_result.push(chunk_result);
+                remaining = new_remaining;
+            }
+            Ok((format!("[{}]", nested_result.join(", ")), remaining))
+        }
+
+        TyKind::Custom {
+            module,
+            name: struct_name,
+        } => {
+            let mut nested_result = Vec::new();
+            let mut remaining = var_info_var.to_vec();
+            for _ in 0..size {
+                let mut string_vec = Vec::new();
+                let (output, new_remaining) = log_custom_type(
+                    backend,
+                    module,
+                    struct_name,
+                    typed,
+                    &remaining,
+                    witness,
+                    span,
+                    &mut string_vec,
+                )?;
+                nested_result.push(format!("{}{}", struct_name, output));
+                remaining = new_remaining;
+            }
+            Ok((format!("[{}]", nested_result.join(", ")), remaining))
+        }
+
+        TyKind::GenericSizedArray(_, _) => {
+            unreachable!("GenericSizedArray should be monomorphized")
+        }
+
+        TyKind::String(_) => todo!("String type cannot be used in circuits!"),
+    }
+}
+pub fn log_custom_type<B: Backend>(
+    backend: &B,
+    module: &ModulePath,
+    struct_name: &String,
+    typed: &Mast<B>,
+    var_info_var: &[ConstOrCell<B::Field, B::Var>],
+    witness: &mut WitnessEnv<B::Field>,
+    span: &Span,
+    string_vec: &mut Vec<String>,
+) -> Result<(String, Vec<ConstOrCell<B::Field, B::Var>>)> {
+    let qualified = FullyQualified::new(module, struct_name);
+    let struct_info = typed
+        .struct_info(&qualified)
+        .ok_or(
+            typed
+                .0
+                .error(ErrorKind::UnexpectedError("struct not found"), *span),
+        )
+        .unwrap();
+
+    let mut remaining = var_info_var.to_vec();
+
+    for (field_name, field_typ) in &struct_info.fields {
+        let len = typed.size_of(field_typ);
+        match field_typ {
+            TyKind::Field { .. } => match &remaining[0] {
+                ConstOrCell::Const(cst) => {
+                    string_vec.push(format!("{field_name}: {}", cst.pretty()));
+                    remaining = remaining[len..].to_vec();
+                }
+                ConstOrCell::Cell(cell) => {
+                    let val = backend.compute_var(witness, cell)?;
+                    string_vec.push(format!("{field_name}: {}", val.pretty()));
+                    remaining = remaining[len..].to_vec();
+                }
+            },
+
+            TyKind::Bool => match &remaining[0] {
+                ConstOrCell::Const(cst) => {
+                    let val = *cst == B::Field::one();
+                    string_vec.push(format!("{field_name}: {}", val));
+                    remaining = remaining[len..].to_vec();
+                }
+                ConstOrCell::Cell(cell) => {
+                    let val = backend.compute_var(witness, cell)? == B::Field::one();
+                    string_vec.push(format!("{field_name}: {}", val));
+                    remaining = remaining[len..].to_vec();
+                }
+            },
+
+            TyKind::Array(b, s) => {
+                let (output, new_remaining) =
+                    log_array_type(backend, &remaining, b, *s, witness, typed, span)?;
+                string_vec.push(format!("{field_name}: {}", output));
+                remaining = new_remaining;
+            }
+
+            TyKind::Custom {
+                module,
+                name: struct_name,
+            } => {
+                let mut custom_string_vec = Vec::new();
+                let (output, new_remaining) = log_custom_type(
+                    backend,
+                    module,
+                    struct_name,
+                    typed,
+                    &remaining,
+                    witness,
+                    span,
+                    &mut custom_string_vec,
+                )?;
+                string_vec.push(format!("{}: {}{}", field_name, struct_name, output));
+                remaining = new_remaining;
+            }
+
+            TyKind::GenericSizedArray(_, _) => {
+                unreachable!("GenericSizedArray should be monomorphized")
+            }
+            TyKind::String(s) => {
+                todo!("String cannot be a type for customs it is only for logging")
+            }
+        }
+    }
+
+    Ok((format!("{{ {} }}", string_vec.join(", ")), remaining))
 }
